+++ src/test/kotlin/lotto/domain/LottoTest.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.domain
 
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
